# ConvToMD: Remove extra calls to joinAll().